das-client/src/main/java/com/ppdai/das/core/datasource/RefreshableDataSource.java
package com.ppdai.das.core.datasource;

import java.io.PrintWriter;
import java.sql.Connection;
import java.sql.SQLException;
import java.sql.SQLFeatureNotSupportedException;
import java.util.concurrent.atomic.AtomicReference;

import javax.sql.DataSource;

import com.ppdai.das.core.configure.DataSourceConfigure;
import com.ppdai.das.core.configure.DataSourceConfigureChangeEvent;
import com.ppdai.das.core.configure.DataSourceConfigureChangeListener;

public class RefreshableDataSource implements DataSource, DataSourceConfigureChangeListener {
    private AtomicReference<SingleDataSource> dataSourceReference = new AtomicReference<>();

    public RefreshableDataSource(String name, DataSourceConfigure config) throws SQLException {
        SingleDataSource dataSource = new SingleDataSource(name, config);
        dataSourceReference.set(dataSource);
    }

    @Override
    public synchronized void configChanged(DataSourceConfigureChangeEvent event) throws SQLException {
        String name = event.getName();
        DataSourceConfigure newConfigure = event.getNewDataSourceConfigure();
        SingleDataSource oldDataSource = dataSourceReference.getAndSet(newConfigure == null ? null : new SingleDataSource(name, newConfigure));
        close(oldDataSource);
    }

    private void close(SingleDataSource oldDataSource) {


        DataSourceTerminator.getInstance().close(oldDataSource);
    }

    private DataSource getDataSource() {
        DataSource dataSource = dataSourceReference.get().getDataSource();
        if (dataSource == null)
            throw new IllegalStateException("DataSource cannot be null.");
        return dataSource;
    }

    @Override
    public Connection getConnection() throws SQLException {
        return getDataSource().getConnection();
    }

    @Override
    public Connection getConnection(String paramString1, String paramString2) throws SQLException {
        return getDataSource().getConnection(paramString1, paramString2);
    }

    @Override
    public PrintWriter getLogWriter() throws SQLException {
        return getDataSource().getLogWriter();
    }

    @Override
    public int getLoginTimeout() throws SQLException {
        return getDataSource().getLoginTimeout();
    }

    @Override
    public void setLogWriter(PrintWriter paramPrintWriter) throws SQLException {
        getDataSource().setLogWriter(paramPrintWriter);
    }

    @Override
    public void setLoginTimeout(int paramInt) throws SQLException {
        getDataSource().setLoginTimeout(paramInt);
    }

    @Override
    public java.util.logging.Logger getParentLogger() throws SQLFeatureNotSupportedException {
        return getDataSource().getParentLogger();
    }

    @Override
    public <T> T unwrap(Class<T> iface) throws SQLException {
        return getDataSource().unwrap(iface);
    }

    @Override
    public boolean isWrapperFor(Class<?> iface) throws SQLException {
        return getDataSource().isWrapperFor(iface);
    }
}
